requirements.txt
ttkbootstrap
Pillow
pyinstaller
